# src/main/resources/document/默认分组_OpenAPI.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "接口文档",
    "contact": {}
  },
  "host": "localhost:8123",
  "basePath": "/api",
  "tags": [
    {
      "name": "测试接口",
      "x-order": "2147483647"
    },
    {
      "name": "用户接口",
      "x-order": "2147483647"
    },
    {
      "name": "课程接口",
      "x-order": "2147483647"
    }
  ],
  "paths": {
    "/api/course": {
      "post": {
        "tags": [
          "课程接口"
        ],
        "summary": "添加课程",
        "operationId": "addCourseUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "courseDTO",
            "description": "courseDTO",
            "required": true,
            "schema": {
              "originalRef": "CourseDTO",
              "$ref": "#/definitions/CourseDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "Result",
              "$ref": "#/definitions/Result"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "Result",
              "$ref": "#/definitions/Result"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      },
      "put": {
        "tags": [
          "课程接口"
        ],
        "summary": "修改课程数据",
        "operationId": "updateCourseUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "courseDTO",
            "description": "courseDTO",
            "required": true,
            "schema": {
              "originalRef": "CourseDTO",
              "$ref": "#/definitions/CourseDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "Result",
              "$ref": "#/definitions/Result"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "Result",
              "$ref": "#/definitions/Result"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/course/{id}": {
      "delete": {
        "tags": [
          "课程接口"
        ],
        "summary": "删除课程",
        "operationId": "deleteCourseUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "Result",
              "$ref": "#/definitions/Result"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "Result",
              "$ref": "#/definitions/Result"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/course/{userId}": {
      "get": {
        "tags": [
          "课程接口"
        ],
        "summary": "获取用户所有课程",
        "operationId": "getCoursesByUserIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "userId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "Result«List«Course»»",
              "$ref": "#/definitions/Result«List«Course»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "Result«List«Course»»",
              "$ref": "#/definitions/Result«List«Course»»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/test": {
      "get": {
        "tags": [
          "测试接口"
        ],
        "summary": "测试接口",
        "operationId": "testUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "Result«string»",
              "$ref": "#/definitions/Result«string»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "Result«string»",
              "$ref": "#/definitions/Result«string»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/user/login": {
      "post": {
        "tags": [
          "用户接口"
        ],
        "summary": "用户登录",
        "operationId": "loginUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "userDTO",
            "description": "userDTO",
            "required": true,
            "schema": {
              "originalRef": "UserDTO",
              "$ref": "#/definitions/UserDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "Result«UserLoginVO»",
              "$ref": "#/definitions/Result«UserLoginVO»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "Result«UserLoginVO»",
              "$ref": "#/definitions/Result«UserLoginVO»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/api/user/register": {
      "post": {
        "tags": [
          "用户接口"
        ],
        "summary": "用户注册",
        "operationId": "registerUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "userDTO",
            "description": "userDTO",
            "required": true,
            "schema": {
              "originalRef": "UserDTO",
              "$ref": "#/definitions/UserDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "Result",
              "$ref": "#/definitions/Result"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "responsesObject": {
          "200": {
            "description": "OK",
            "schema": {
              "originalRef": "Result",
              "$ref": "#/definitions/Result"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    }
  },
  "definitions": {
    "Course": {
      "type": "object",
      "properties": {
        "courseName": {
          "type": "string",
          "refType": null
        },
        "courseType": {
          "type": "integer",
          "format": "int32",
          "refType": null
        },
        "credits": {
          "type": "integer",
          "format": "int32",
          "refType": null
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "totalHours": {
          "type": "integer",
          "format": "int32",
          "refType": null
        },
        "userId": {
          "type": "integer",
          "format": "int64",
          "refType": null
        }
      },
      "title": "Course"
    },
    "CourseDTO": {
      "type": "object",
      "properties": {
        "courseName": {
          "type": "string",
          "refType": null
        },
        "courseType": {
          "type": "integer",
          "format": "int32",
          "refType": null
        },
        "credits": {
          "type": "integer",
          "format": "int32",
          "refType": null
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "refType": null
        },
        "preRequisiteCourseIds": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int64"
          },
          "refType": "integer"
        },
        "totalHours": {
          "type": "integer",
          "format": "int32",
          "refType": null
        }
      },
      "title": "CourseDTO"
    },
    "Result": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "object"
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "Result"
    },
    "Result«List«Course»»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "array",
          "items": {
            "originalRef": "Course",
            "$ref": "#/definitions/Course"
          }
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "Result«List«Course»»"
    },
    "Result«UserLoginVO»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "originalRef": "UserLoginVO",
          "$ref": "#/definitions/UserLoginVO"
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "Result«UserLoginVO»"
    },
    "Result«string»": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "string"
        },
        "msg": {
          "type": "string"
        }
      },
      "title": "Result«string»"
    },
    "UserDTO": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string",
          "refType": null
        },
        "username": {
          "type": "string",
          "refType": null
        }
      },
      "title": "UserDTO"
    },
    "UserLoginVO": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "description": "id",
          "refType": null
        },
        "token": {
          "type": "string",
          "description": "token",
          "refType": null
        },
        "username": {
          "type": "string",
          "description": "用户名",
          "refType": null
        }
      },
      "title": "UserLoginVO",
      "description": "给前端返回的登录信息"
    }
  },
  "x-openapi": {
    "x-markdownFiles": null,
    "x-setting": {
      "language": "zh-CN",
      "enableSwaggerModels": true,
      "swaggerModelName": "Swagger Models",
      "enableReloadCacheParameter": false,
      "enableAfterScript": true,
      "enableDocumentManage": true,
      "enableVersion": false,
      "enableRequestCache": true,
      "enableFilterMultipartApis": false,
      "enableFilterMultipartApiMethodType": "POST",
      "enableHost": false,
      "enableHostText": "",
      "enableDynamicParameter": false,
      "enableDebug": true,
      "enableFooter": true,
      "enableFooterCustom": false,
      "footerCustomContent": null,
      "enableSearch": true,
      "enableOpenApi": true,
      "enableHomeCustom": false,
      "homeCustomLocation": null,
      "enableGroup": true,
      "enableResponseCode": true
    }
  }
}